Correction to task code
Brief: This code is meant to update the older PRESSB code. There was an issue that was effectively doubling the number of responses.

Extended: The number of responses was effectively doubled due to how VB.Net buttons functions. To create the 'depressable' button we had inherited from the Button class, which in turn inherits from the Control class. At the Button level, there is a separation of the keyDown (when the key is depressed) and keyUp (when the key is released) functions, but when you go down a level to the Control class it becomes clear that both call the same function. Since I had written the button class to call the super functions this resulted in keyUp having an effect on the behavioral count (because it called the same function as keyDown) when it should not have. It is generally easy to determine the keyUp behavior from the keyDown behavior because the interval between the recorded presses is quite different (generally ~ 100 ms on the keyUp counts). Regardless time allocation is properly recorded since each response's ms time point is recorded.